requirements.txt
Flask==2.3.3
Flask-Caching==2.0.2
python-dotenv==1.0.0
requests==2.31.0
pdfkit==1.0.0
Werkzeug==2.3.7
gunicorn==21.2.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.0.5
psycopg2-binary==2.9.7
WeasyPrint==59.0
Pillow==10.0.1
matplotlib==3.10.0
seaborn==0.13.2
pandas==2.3.1